[move]ExpTableクラスをlogbook.constantsパッケージへ移動
